netplay: factor `WZFile` out of `netplay.h/cpp`
